# src/routing/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"IErrorResponseDto": {
				"description": "Тело любого ответа с ошибкой (`status >= 400`).",
				"properties": {
					"status": {
						"type": "number",
						"format": "double",
						"description": "HTTP-статус."
					},
					"code": {
						"type": "string",
						"description": "Машинный код: доменный (`USER_EMAIL_TAKEN`) или по статусу (`NOT_FOUND`,\n`VALIDATION_ERROR`, `INTERNAL_ERROR`). Стабилен — на него опирается клиент."
					},
					"message": {
						"type": "string",
						"description": "Человекочитаемое сообщение; для 5xx — общее, без внутренних деталей."
					},
					"details": {
						"description": "Подробности: для VALIDATION_ERROR — поле → сообщение."
					},
					"requestId": {
						"type": "string",
						"description": "Идентификатор запроса — для обращения в поддержку и поиска в логах."
					},
					"stack": {
						"type": "string",
						"description": "Stack trace — только вне production."
					}
				},
				"required": [
					"status",
					"code",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UUID": {
				"type": "string",
				"description": "Идентификатор сущности в пути или теле запроса. tsoa проверяет формат до\nконтроллера: неверный uuid → 422, а не ошибка Postgres.",
				"pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$"
			},
...
			"EWgEndpointMode": {
				"description": "Режим точки подключения.",
				"enum": [
					"direct",
					"relay"
				],
				"type": "string"
			},
			"EWgForwardMode": {
				"description": "Как релей гонит трафик до целевой ноды.",
				"enum": [
					"dnat",
					"ipip"
				],
				"type": "string"
			},
			"WgEndpointDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"host": {
						"type": "string"
					},
					"mode": {
						"$ref": "#/components/schemas/EWgEndpointMode"
					},
					"relayNodeId": {
						"type": "string",
						"nullable": true
					},
					"forwardMode": {
						"$ref": "#/components/schemas/EWgForwardMode"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"host",
					"mode",
					"relayNodeId",
					"forwardMode",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateWgEndpointBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"host": {
						"type": "string",
						"description": "Хост (IP/домен), который попадает в клиентские конфиги."
					},
					"mode": {
						"$ref": "#/components/schemas/EWgEndpointMode"
					},
					"relayNodeId": {
						"type": "string",
						"nullable": true,
						"description": "Обязательна для mode=relay."
					},
					"forwardMode": {
						"$ref": "#/components/schemas/EWgForwardMode"
					}
				},
				"required": [
					"name",
					"host",
					"mode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPaginatedDto_WgEndpointDto_": {
				"description": "Страница по смещению: списки с известным общим числом.",
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/WgEndpointDto"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double",
						"description": "Всего элементов, подходящих под фильтр."
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"total",
					"offset",
					"limit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WgEndpointOptionDto": {
				"description": "Краткая запись для выпадающих списков.",
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"host": {
						"type": "string"
					},
					"mode": {
						"$ref": "#/components/schemas/EWgEndpointMode"
					}
				},
				"required": [
					"id",
					"name",
					"host",
					"mode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUpdateWgEndpointBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"host": {
						"type": "string"
					},
					"mode": {
						"$ref": "#/components/schemas/EWgEndpointMode"
					},
					"relayNodeId": {
						"type": "string",
						"nullable": true
					},
					"forwardMode": {
						"$ref": "#/components/schemas/EWgForwardMode"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"EWgForwardProtocol": {
				"enum": [
					"udp",
					"tcp"
				],
				"type": "string"
			},
			"EWgForwardPath": {
				"description": "Путь до цели: напрямую (DNAT на её адрес) или через IPIP-туннель.",
				"enum": [
					"direct",
					"ipip"
				],
				"type": "string"
			},
			"EWgForwardRoute": {
				"description": "Маршрут пути `ipip`: `auto` — туннель, а если он не отвечает, напрямую;\n`tunnel` / `direct` — принудительно.",
				"enum": [
					"auto",
					"tunnel",
					"direct"
				],
				"type": "string"
			},
			"EWgForwardActiveRoute": {
				"description": "Активный маршрут проброса по отчёту агента релея.",
				"enum": [
					"tunnel",
					"direct"
				],
				"type": "string"
			},
			"WgForwardDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"relayNodeId": {
						"type": "string"
					},
					"relayNodeName": {
						"type": "string",
						"nullable": true
					},
					"protocol": {
						"$ref": "#/components/schemas/EWgForwardProtocol"
					},
					"listenPort": {
						"type": "number",
						"format": "double"
					},
					"targetNodeId": {
						"type": "string",
						"nullable": true
					},
					"targetNodeName": {
						"type": "string",
						"nullable": true
					},
					"targetHost": {
						"type": "string",
						"nullable": true
					},
					"targetPort": {
						"type": "number",
						"format": "double"
					},
					"path": {
						"$ref": "#/components/schemas/EWgForwardPath"
					},
					"route": {
						"$ref": "#/components/schemas/EWgForwardRoute"
					},
					"enabled": {
						"type": "boolean"
					},
					"activeRoute": {
						"allOf": [
							{
								"$ref": "#/components/schemas/EWgForwardActiveRoute"
							}
						],
						"nullable": true,
						"description": "Маршрут по последнему отчёту агента релея; null — отчёта ещё нет."
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"relayNodeId",
					"relayNodeName",
					"protocol",
					"listenPort",
					"targetNodeId",
					"targetNodeName",
					"targetHost",
					"targetPort",
					"path",
					"route",
					"enabled",
					"activeRoute",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICreateWgForwardBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"relayNodeId": {
						"type": "string"
					},
					"protocol": {
						"$ref": "#/components/schemas/EWgForwardProtocol"
					},
					"listenPort": {
						"type": "number",
						"format": "double"
					},
					"targetNodeId": {
						"type": "string",
						"nullable": true,
						"description": "Нода-цель с агентом (обязательна для пути `ipip`)."
					},
					"targetHost": {
						"type": "string",
						"nullable": true,
						"description": "Прямой адрес цели; пусто — publicHost ноды-цели."
					},
					"targetPort": {
						"type": "number",
						"format": "double"
					},
					"path": {
						"$ref": "#/components/schemas/EWgForwardPath"
					},
					"route": {
						"$ref": "#/components/schemas/EWgForwardRoute"
					},
					"enabled": {
						"type": "boolean"
					}
				},
				"required": [
					"name",
					"relayNodeId",
					"protocol",
					"listenPort",
					"targetPort",
					"path"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPaginatedDto_WgForwardDto_": {
				"description": "Страница по смещению: списки с известным общим числом.",
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/WgForwardDto"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double",
						"description": "Всего элементов, подходящих под фильтр."
					},
					"offset": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"items",
					"total",
					"offset",
					"limit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUpdateWgForwardBody": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"listenPort": {
						"type": "number",
						"format": "double"
					},
					"targetNodeId": {
						"type": "string",
						"nullable": true
					},
					"targetHost": {
						"type": "string",
						"nullable": true
					},
					"targetPort": {
						"type": "number",
						"format": "double"
					},
					"path": {
						"$ref": "#/components/schemas/EWgForwardPath"
					},
					"route": {
						"$ref": "#/components/schemas/EWgForwardRoute"
					},
					"enabled": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"apiKey": {
				"type": "apiKey",
				"in": "header",
				"name": "X-Api-Key",
				"description": "API-ключ: агент ноды (scope wg-agent) или интеграция"
			}
		}
	},
	"info": {
		"title": "wg-admin",
		"version": "1.0.1",
		"description": "WG Admin — админка WireGuard: бэкенд (Koa, tsoa, TypeORM, pg-boss) и агент нод на Go",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Epifanov Andrei",
			"email": "epifanovmd@gmail.com"
		}
	},
	"paths": {
...
		"/api/v1/wg/endpoints": {
			"post": {
				"operationId": "CreateWgEndpoint",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WgEndpointDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Создать точку подключения — стабильный адрес для клиентских конфигов\n(напрямую или через релей-ноду).",
				"summary": "Создание точки подключения",
				"tags": [
					"WgEndpoint"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:endpoint:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateWgEndpointBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListWgEndpoints",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaginatedDto_WgEndpointDto_"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Точки подключения, новые первыми.",
				"summary": "Список точек подключения",
				"tags": [
					"WgEndpoint"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:endpoint:view"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/wg/endpoints/options": {
			"get": {
				"operationId": "WgEndpointOptions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/WgEndpointOptionDto"
									},
									"type": "array"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Краткий список для выпадающих списков.",
				"summary": "Точки подключения (options)",
				"tags": [
					"WgEndpoint"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:endpoint:view"
						]
					}
				],
				"parameters": []
			}
		},
		"/api/v1/wg/endpoints/{id}": {
			"get": {
				"operationId": "GetWgEndpoint",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WgEndpointDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Точка подключения по id.",
				"summary": "Точка подключения",
				"tags": [
					"WgEndpoint"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:endpoint:view"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateWgEndpoint",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WgEndpointDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Изменить точку подключения; смена хоста/релея применяется к нодам\nавтоматически, клиентские конфиги перевыпускать не нужно.",
				"summary": "Изменение точки подключения",
				"tags": [
					"WgEndpoint"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:endpoint:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUpdateWgEndpointBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteWgEndpoint",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Удалить точку подключения; используемая интерфейсами — 409.",
				"summary": "Удаление точки подключения",
				"tags": [
					"WgEndpoint"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:endpoint:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
...
		"/api/v1/wg/forwards": {
			"post": {
				"operationId": "CreateWgForward",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WgForwardDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Создать проброс: релей, протокол и порт, цель (нода с агентом или\nадрес), путь и режим маршрута.",
				"summary": "Создание проброса",
				"tags": [
					"WgForward"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:forward:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ICreateWgForwardBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ListWgForwards",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IPaginatedDto_WgForwardDto_"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Пробросы с активным маршрутом по отчётам агентов.",
				"summary": "Список пробросов",
				"tags": [
					"WgForward"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:forward:view"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/wg/forwards/{id}": {
			"get": {
				"operationId": "GetWgForward",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WgForwardDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"summary": "Проброс",
				"tags": [
					"WgForward"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:forward:view"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateWgForward",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WgForwardDto"
								}
							}
						}
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"description": "Изменить проброс; в том числе переключить маршрут (auto / tunnel /\ndirect) — агент релея применит сразу.",
				"summary": "Изменение проброса",
				"tags": [
					"WgForward"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:forward:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUpdateWgForwardBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteWgForward",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"default": {
						"description": "Ошибка",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponseDto"
								}
							}
						}
					}
				},
				"summary": "Удаление проброса",
				"tags": [
					"WgForward"
				],
				"security": [
					{
						"jwt": [
							"permission:wg:forward:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8181",
			"description": "Local dev"
		}
	]
}